\name{EU_stock_markets}
\alias{EU_stock_markets}
\title{Financial data set to exemplify the use of the functions contained in 'RolWinWavCor'
}
\description{
The data set \code{\link{EU_stock_markets}} contains nine European stock market indices (closing prices): PSI20 (Portugal), MIB30 (Italy), ISEQ (Ireland), ASE20 (Greece), IBEX35 (Spain), DAX30 (Germany), CAC40 (France) and SPEurope350 (Eurozone). The first column are the dates and these have the format YYYY-MM-DD. All data sets come from Yahoo Finance (except for the PSI20 and SPEurope350 stock market indices, which were taken from BolsaPT and OnVista Group) and encompass the interval from January 2, 2004 to December 31, 2011 (2086 elements or data points). This data set was used in Polanco-Martínez et al. (2018) and this work should be consulted for more information about this financial data set. 
}
\usage{data(EU_stock_markets)
}
\format{Data set is in ASCII format containing 10 columns and 2086 rows (elements or data points), columns are separated by spaces. 
}
\source{
Polanco-Martínez, J. M., Fernández-Macho, J., Neumann, M. B., & Faria, S. H. (2018). A pre-crisis vs. crisis analysis of peripheral EU stock markets by means of wavelet transform and a nonlinear causality test. Physica A: Statistical Mechanics and its Applications, 490, 1211-1227. <URL: \doi{10.1016/j.physa.2017.08.065}>. 
}
\keyword{EU_stock_markets}
\keyword{dataset_EU_stock_markets}
